# [Cuda] Refactor GroupNorm