--- a/onnxruntime/contrib_ops/cuda/diffusion/group_norm_impl_kernel.cuh
+++ b/onnxruntime/contrib_ops/cuda/diffusion/group_norm_impl_kernel.cuh
@@ -19,10 +19,8 @@
 // Modifications: heuristic channels per block; support epsilon; support skip and bias; update coding style.
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
+#pragma once
 #include <cuda_fp16.h>
-#include <cuda_runtime_api.h>
-#include <cub/cub.cuh>
 #include "core/providers/cuda/cuda_common.h"
 #include "core/providers/cuda/cu_inc/common.cuh"