… cmake/onnxruntime_rocm_hipify.cmake
… onnxruntime/contrib_ops/cuda/diffusion/group_norm_impl.cu