Split GroupNorm implementation into multiple files, to make ROCm EP can reuse cuda code.

Related PR: https://github.com/microsoft/onnxruntime/pull/19158