@PeixuanZuo 2024-01-23
Run script onnxruntime/test/python/transformers/test_group_norm.py, compare performance between main branch and [peixuanzuo/refactor_groupnorm](https://github.com/microsoft/onnxruntime/tree/peixuanzuo/refactor_groupnorm) branch.

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/PEIXUA~1/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/PEIXUA~1/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<style>
<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
.xl65
	{mso-number-format:Percent;}
.xl66
	{mso-number-format:Percent;
	background:#EDEDED;
	mso-pattern:black none;}
-->
</style>
</head>

<body link="#0563C1" vlink="#954F72">


Configuration | Latency(us)    branch main | Latency(us)    branch refactor1 | gap
-- | -- | -- | --
B: 2  H: 64  W: 64  C: 128    G: 32  activation: 0  channels_last: 1  fp16: 1 | 69 | 67 | 2.90%
B: 2  H: 64  W: 64  C: 256    G: 32  activation: 0  channels_last: 1  fp16: 1 | 79 | 76 | 3.80%
B: 2  H: 64  W: 64  C: 512    G: 32  activation: 0  channels_last: 1  fp16: 1 | 95 | 94 | 1.05%
B: 2  H: 64  W: 64  C:   1024  G: 32  activation: 0  channels_last: 1  fp16: 1 | 127 | 127 | 0.00%
B: 2  H: 64  W: 64  C:   2048  G: 32  activation: 0  channels_last: 1  fp16: 1 | 184 | 188 | -2.17%
B: 2  H: 64  W: 64  C: 320    G: 32  activation: 0  channels_last: 1  fp16: 1 | 81 | 80 | 1.23%
B: 2  H: 64  W: 64  C: 640    G: 32  activation: 0  channels_last: 1  fp16: 1 | 103 | 102 | 0.97%
B: 2  H: 64  W: 64  C: 960    G: 32  activation: 0  channels_last: 1  fp16: 1 | 125 | 125 | 0.00%
B: 2  H: 64  W: 64  C:   1920  G: 32  activation: 0  channels_last: 1  fp16: 1 | 179 | 178 | 0.56%
B: 2  H: 64  W: 64  C:   2560  G: 32  activation: 0  channels_last: 1  fp16: 1 | 217 | 219 | -0.92%
B: 2  H: 64  W: 64  C: 384    G: 32  activation: 0  channels_last: 1  fp16: 1 | 85 | 83 | 2.35%
B: 2  H: 64  W: 64  C: 768    G: 32  activation: 0  channels_last: 1  fp16: 1 | 119 | 107 | 10.08%
B: 2  H: 64  W: 64  C:   1536  G: 32  activation: 0  channels_last: 1  fp16: 1 | 166 | 153 | 7.83%
B: 2  H: 64  W: 64  C:   3072  G: 32  activation: 0  channels_last: 1  fp16: 1 | 251 | 249 | 0.80%
B: 2  H: 64  W: 64  C:   1152  G: 32  activation: 0  channels_last: 1  fp16: 1 | 137 | 135 | 1.46%
B: 2  H: 64  W: 64  C:   2304  G: 32  activation: 0  channels_last: 1  fp16: 1 | 221 | 231 | -4.52%
B: 2  H: 96  W: 96  C: 128    G: 32  activation: 0  channels_last: 1  fp16: 1 | 86 | 83 | 3.49%
B: 2  H: 96  W: 96  C: 256    G: 32  activation: 0  channels_last: 1  fp16: 1 | 114 | 104 | 8.77%
B: 2  H: 96  W: 96  C: 512    G: 32  activation: 0  channels_last: 1  fp16: 1 | 135 | 133 | 1.48%
B: 2  H: 96  W: 96  C:   1024  G: 32  activation: 0  channels_last: 1  fp16: 1 | 204 | 201 | 1.47%
B: 2  H: 96  W: 96  C:   2048  G: 32  activation: 0  channels_last: 1  fp16: 1 | 346 | 344 | 0.58%
B: 2  H: 96  W: 96  C: 320    G: 32  activation: 0  channels_last: 1  fp16: 1 | 115 | 114 | 0.87%
B: 2  H: 96  W: 96  C: 640    G: 32  activation: 0  channels_last: 1  fp16: 1 | 163 | 155 | 4.91%
B: 2  H: 96  W: 96  C: 960    G: 32  activation: 0  channels_last: 1  fp16: 1 | 207 | 207 | 0.00%
B: 2  H: 96  W: 96  C:   1920  G: 32  activation: 0  channels_last: 1  fp16: 1 | 340 | 327 | 3.82%
B: 2  H: 96  W: 96  C:   2560  G: 32  activation: 0  channels_last: 1  fp16: 1 | 429 | 417 | 2.80%
B: 2  H: 96  W: 96  C: 384    G: 32  activation: 0  channels_last: 1  fp16: 1 | 121 | 118 | 2.48%
B: 2  H: 96  W: 96  C: 768    G: 32  activation: 0  channels_last: 1  fp16: 1 | 172 | 174 | -1.16%
B: 2  H: 96  W: 96  C:   1536  G: 32  activation: 0  channels_last: 1  fp16: 1 | 298 | 276 | 7.38%
B: 2  H: 96  W: 96  C:   3072  G: 32  activation: 0  channels_last: 1  fp16: 1 | 498 | 501 | -0.60%
B: 2  H: 96  W: 96  C:   1152  G: 32  activation: 0  channels_last: 1  fp16: 1 | 223 | 221 | 0.90%
B: 2  H: 96  W: 96  C:   2304  G: 32  activation: 0  channels_last: 1  fp16: 1 | 397 | 398 | -0.25%
B: 2  H: 128  W: 128  C:   128  G: 32  activation: 0  channels_last: 1  fp16: 1 | 103 | 102 | 0.97%
B: 2  H: 128  W: 128  C:   256  G: 32  activation: 0  channels_last: 1  fp16: 1 | 136 | 127 | 6.62%
B: 2  H: 128  W: 128  C:   512  G: 32  activation: 0  channels_last: 1  fp16: 1 | 202 | 195 | 3.47%
B: 2  H: 128  W: 128  C:   1024  G: 32  activation: 0  channels_last: 1  fp16: 1 | 324 | 327 | -0.93%
B: 2  H: 128  W: 128  C:   2048  G: 32  activation: 0  channels_last: 1  fp16: 1 | 574 | 588 | -2.44%
B: 2  H: 128  W: 128  C:   320  G: 32  activation: 0  channels_last: 1  fp16: 1 | 156 | 150 | 3.85%
B: 2  H: 128  W: 128  C:   640  G: 32  activation: 0  channels_last: 1  fp16: 1 | 230 | 228 | 0.87%
B: 2  H: 128  W: 128  C:   960  G: 32  activation: 0  channels_last: 1  fp16: 1 | 331 | 327 | 1.21%
B: 2  H: 128  W: 128  C:   1920  G: 32  activation: 0  channels_last: 1  fp16: 1 | 549 | 543 | 1.09%
B: 2  H: 128  W: 128  C:   2560  G: 32  activation: 0  channels_last: 1  fp16: 1 | 697 | 697 | 0.00%
B: 2  H: 128  W: 128  C:   384  G: 32  activation: 0  channels_last: 1  fp16: 1 | 179 | 179 | 0.00%
B: 2  H: 128  W: 128  C:   768  G: 32  activation: 0  channels_last: 1  fp16: 1 | 263 | 259 | 1.52%
B: 2  H: 128  W: 128  C:   1536  G: 32  activation: 0  channels_last: 1  fp16: 1 | 463 | 462 | 0.22%
B: 2  H: 128  W: 128  C:   3072  G: 32  activation: 0  channels_last: 1  fp16: 1 | 825 | 829 | -0.48%
B: 2  H: 128  W: 128  C:   1152  G: 32  activation: 0  channels_last: 1  fp16: 1 | 378 | 360 | 4.76%
B: 2  H: 128  W: 128  C:   2304  G: 32  activation: 0  channels_last: 1  fp16: 1 | 687 | 681 | 0.87%



</body>

</html>