@PeixuanZuo 2024-01-23
> @PeixuanZuo, Could you run onnxruntime/test/python/transformers/test_group_norm.py (It is not enabled in pipeline) to see whether there is regression in performance/parity?
> 
> You can run the script before/after your change, and manually compare the latency numbers in Console output. Maybe add the results in the description of this PR

Thanks @tianleiwu, I run onnxruntime/test/python/transformers/test_group_norm.py multiple times, the performance results fluctuate each time. I pasted the last test result here.